refactor(notify): implement service worker notification handler and refactor main file
